int -> long

+++ walking-beans-backend/src/main/java/walking_beans/walking_beans_backend/model/dto/Address.java
@@ -14,12 +14,18 @@ public class Address {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int addressId;
-    private int userId;
+    private long addressId;
+
+    private long userId;
+
     private String address;
+
     private String addressName;
+
     private double addressLatitude;
+
     private double addressLongitude;
-    private int addressRole;
+
+    private byte addressRole;
 
 }
